Local hook template, demo cal file, demo works again.
OLFlickerDemo was broken, because we no longer have a ‘OneLight’ preference group.  I added a OneLightToolbox local hook which sets a calibration file preference, and I added an OLDemoCal folder to the toolbox which has OLDemoCal.m in it.  So now the demo works again.

When we access calibration files from other approaches, we should use the calibration file preference for that approach to point to where the cal files live.  You do this by passing the ‘CalibrationFolder’ key/value pair to the routines that find calibration files, and using the preference to point at the folder.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>