Finished off support for modulations, version 0.9
There were hooks that looked like things were mostly set up for buidling nominal primaries for bipolar modulations, but it wasn’t done.  Got it to work, mostly by adding ‘modulation’ to various switch statements.

Calling it version 0.9 because it’s not particularly tested.

Fixed a bug introduced in check for primary range, now we think that is working.