Renamed directions: pulse -> unipolar & modulation -> bipolar
Directions of modulation have been called ‘pulse’ and ‘modulation’. However, these are descriptions of the temporal aspects usually associated with them. The practical difference between these directions, is that pulses are usually done with unipolar contrast, and more complex, e.g., sinusoidal, modulations are usually done with bipolar contrast. The renaming should clarify the spectral difference between these directions.

Moreover, ‘modulation’ was used to describe a combination of spectral and temporal variation, but also to describe a specific kind of spectral variation. The renaming should clarify this distinction between the spectral, and temporal, aspects of a modulation.