Removed “predicted” excitations/contrasts
The “predicted” excitatons/contrasts were based on the SPD we would predict from the primary values  and calibration. This is (likely) the same as desired SPD/excitations/contrasts for nominal (uncorrected) directions, but not for corrected directions (which break the link between calibration and primaries).

This is confusing (predicted from primary values is not the same as simulated, or desired), and likely useless: probably no case in which we want to know what we would predict using a calibration we know is not correct….