Pass direction age to routine that makes background
We used to always make the background for 32 year olds, and then construct directions for all ages.  The reason we didn’t make backgrounds for all ages is that it took a long time, and we then cached the primaries for all ages.

But now we have ditched the caching, and we make the background specifically for each observer.  So, it makes sense to use the actual observer’s age.

If the routine that makes directions from parameters also makes the background, and if a single age (rather than a list of ages) is passed to that routine, then it then uses the same age to make the background primaries, overriding what is in the background parameters structure.

If background primaries are made separately, then they get made with whatever age is in the background parameter structure.

An alternative way to handle this is to remove the age parameter from the background params object, and have a key/value pair that sets the age when the background gets made.  This would parallel the way it is done for directions and be cleaner.  Not doing it right now, just noting.