Hotfix: `OLDirection.SPDdifferentialDesired` would not scale correctly
Multiplying a direction by a scalar would correctly scale the differential primary values, but then calculate the SPD differentialDesired from the new primary values. In the case of a corrected direction, where the primary values and SPD are no longer a linear transformation, this would yield an incorrect `SPDdifferentialDesired`.
This fix overwrites the `SPDdifferentialDesired` of the new, scaled, direction, with a scaled version of the original SPD.